+++ src/features/script-load.js
@@ -1,27 +1,67 @@
 /*
- * Supports loading System.register via script tag injection
+ * Script instantiation loading
  */
-
-import { systemJSPrototype } from '../system-core';
-import { hasDocument, baseUrl, resolveUrl } from '../common';
+import { hasDocument } from '../common.js';
+import { systemJSPrototype } from '../system-core.js';
 import { errMsg } from '../err-msg.js';
+import { importMap } from './import-maps.js';
 
-var systemRegister = systemJSPrototype.register;
-systemJSPrototype.register = function (deps, declare) {
-  systemRegister.call(this, deps, declare);
-};
+if (hasDocument) {
+  window.addEventListener('error', function (evt) {
+    lastWindowErrorUrl = evt.filename;
+    lastWindowError = evt.error;
+  });
+  var baseOrigin = location.origin;
+}
 
 systemJSPrototype.createScript = function (url) {
   var script = document.createElement('script');
-  script.charset = 'utf-8';
   script.async = true;
-  script.crossOrigin = 'anonymous';
+  // Only add cross origin for actual cross origin
+  // this is because Safari triggers for all
+  // - https://bugs.webkit.org/show_bug.cgi?id=171566
+  if (url.indexOf(baseOrigin + '/'))
+    script.crossOrigin = 'anonymous';
+  var integrity = importMap.integrity[url];
+  if (integrity)
+    script.integrity = integrity;
   script.src = url;
   return script;
 };
 
+// Auto imports -> script tags can be inlined directly for load phase
+var lastAutoImportUrl, lastAutoImportDeps, lastAutoImportTimeout;
+var autoImportCandidates = {};
+var systemRegister = systemJSPrototype.register;
+systemJSPrototype.register = function (deps, declare) {
+  if (hasDocument && document.readyState === 'loading' && typeof deps !== 'string') {
+    var scripts = document.querySelectorAll('script[src]');
+    var lastScript = scripts[scripts.length - 1];
+    if (lastScript) {
+      lastAutoImportUrl = lastScript.src;
+      lastAutoImportDeps = deps;
+      // if this is already a System load, then the instantiate has already begun
+      // so this re-import has no consequence
+      var loader = this;
+      lastAutoImportTimeout = setTimeout(function () {
+        autoImportCandidates[lastScript.src] = [deps, declare];
+        loader.import(lastScript.src);
+      });
+    }
+  }
+  else {
+    lastAutoImportDeps = undefined;
+  }
+  return systemRegister.call(this, deps, declare);
+};
+
 var lastWindowErrorUrl, lastWindowError;
 systemJSPrototype.instantiate = function (url, firstParentUrl) {
+  var autoImportRegistration = autoImportCandidates[url];
+  if (autoImportRegistration) {
+    delete autoImportCandidates[url];
+    return autoImportRegistration;
+  }
   var loader = this;
   return new Promise(function (resolve, reject) {
     var script = systemJSPrototype.createScript(url);
@@ -36,29 +76,13 @@ systemJSPrototype.instantiate = function (url, firstParentUrl) {
         reject(lastWindowError);
       }
       else {
-        resolve(loader.getRegister());
+        var register = loader.getRegister();
+        // Clear any auto import registration for dynamic import scripts during load
+        if (register && register[0] === lastAutoImportDeps)
+          clearTimeout(lastAutoImportTimeout);
+        resolve(register);
       }
     });
     document.head.appendChild(script);
   });
 };
-
-if (hasDocument) {
-  window.addEventListener('error', function (evt) {
-    lastWindowErrorUrl = evt.filename;
-    lastWindowError = evt.error;
-  });
-
-  window.addEventListener('DOMContentLoaded', loadScriptModules);
-  loadScriptModules();
-}
-
-
-function loadScriptModules() {
-  [].forEach.call(
-    document.querySelectorAll('script[type=systemjs-module]'), function (script) {
-      if (script.src) {
-        System.import(script.src.slice(0, 7) === 'import:' ? script.src.slice(7) : resolveUrl(script.src, baseUrl));
-      }
-    });
-}