+++ src/system-node.js
@@ -1,12 +1,19 @@
-import { REGISTRY, systemJSPrototype } from './system-core.js';
-import { IMPORT_MAP, IMPORT_MAP_PROMISE } from './features/import-map.js';
+import './features/resolve.js';
 import './features/registry.js';
-import './extras/global.js';
-import './extras/module-types.js';
+import './features/fetch-load.js';
 import './features/node-fetch.js';
-import { BASE_URL, baseUrl, resolveAndComposeImportMap } from './common.js';
+import './extras/global.js';
+
+import { REGISTRY, systemJSPrototype } from './system-core.js';
+import { BASE_URL, baseUrl, resolveAndComposeImportMap, IMPORT_MAP } from './common.js';
 
 export const System = global.System;
+
+const IMPORT_MAP_PROMISE = Symbol();
+
+systemJSPrototype.prepareImport = function () {
+  return this[IMPORT_MAP_PROMISE];
+};
 
 const originalResolve = systemJSPrototype.resolve;
 systemJSPrototype.resolve = function () {
@@ -19,7 +26,8 @@ systemJSPrototype.resolve = function () {
 
 export function applyImportMap(loader, newMap, mapBase) {
   ensureValidSystemLoader(loader);
-  loader[IMPORT_MAP] = resolveAndComposeImportMap(newMap, mapBase || baseUrl, loader[IMPORT_MAP] || { imports: {}, scopes: {} });
+  loader[IMPORT_MAP] = loader[IMPORT_MAP] || { imports: {}, scopes: {} };
+  resolveAndComposeImportMap(newMap, mapBase || baseUrl, loader[IMPORT_MAP]);
   loader[IMPORT_MAP_PROMISE] = Promise.resolve();
 }