+++ test/fixtures/browser/link-error.js
@@ -1,0 +1,6 @@
+System.register(['./link-error-child.js', './link-error-child2.js'], function () {
+  return {
+    execute: function () {
+    }
+  };
+});